Add focus traps to modals
Also, switch to a maintained fork of html-minifier, because something
seems wrong with the unmaintained one. It seemed to be spewing out NULL
bytes?

This change introduces focus traps in modals. This should improve
keyboard accessibility a fair bit, although behavior with nested modals
currently seems to be unreliable, so that might need more investigation.
Focus traps force tab-based focus cycling to remain confined to some
subset of elements, in this case, whichever modal is currently the
topmost. Before this change, the behavior of focus inside modals was
very annoying, making navigating via keyboard not very fun.