Add Open in Photopea item
We can take advantage of the fact that Photopea is purely client-side,
so it can access the hydrus network client just like we can. The only
thing that is terribly sketchy here is the API key being passed through,
since it can administer the hydrus network server. It might be worth
adding the ability to specify a limited API key that can be used to pass
files to services like Photopea... Or maybe hydrus network can add
support for some kind of signed URL scheme, so we can generate URLs that
Photopea can use. not sure. For now, just hide the API key by rewriting
the name using Photopea's built-in scripting capabilities.