Update zip.js + fix workers in Server Mode
It's broken because the default way workers get loaded is through inline
script loading. Inline script loading is effectively an arbitrary code
execution avenue, so we need to disallow it. Thank goodness, less than
24 hours ago, someone actually pushed a fix to zip.js so that it's
possible to import the web worker and wasm modules directly and avoid
the CSP issue, so we can continue to disallow arbitrary inline JS.