Add internal interface + health check endpoint
We could limit it based on where the connections are coming from, but
that would probably make it too easy to accidentally leak the endpoint.
I think people who care about this feature will probably not mind having
it on a separate interface, and this separate interface will probably
also be a good place to put things like Prometheus metrics exporting.

Closes #21.