Enable noUncheckedIndexedAccess (#33)
* Enable noUncheckedIndexedAccess

This is digging into the deeper TypeScript type checking features.
Unfortunately, TypeScript has less power to reason about the correctness
of indexed accesses, so we need to help the type checker a lot in order
to make it happy. Moreover, there are many cases where it is either too
cumbersome to add additional checks that have very little value (like
matrix math, where it would be nice to check bounds but the indices and
array sizes are static so we stand very little to gain, we just want to
crash if the code is wrong) or are performance sensitive (like bitmap
operations; and again, we don't really have much sensible to do here, we
really do need to crash if there's an error in the index calculation for
bitmap operations.)

* Fix autotag bug caused by bad indexing

* Fix overeager missing confidence warning

* Fix existing count not considering rating tags