Improve escape key behavior in input boxes (#37)
When using keyboard navigation, input boxes temporarily disable hotkeys;
this ensures that native shortcuts like Ctrl+A always take precedence
over "global" hotkeys elsewhere in Hydrui. There is one case that needs
particular attention, though: the escape key. Global escape key
shortcuts are used to leave modals, but they're also used to dismiss
autocomplete suggestions and cancel inputs. So, it's still wise to *not*
handle global escape key shortcuts when the focus is on an input widget.
That said, if the user presses escape and it does nothing at all, it
would be useful if it "blurs" the focus. So, if you have autocomplete
suggestions open inside of a modal, the first escape key press closes
the autocomplete suggestions without blurring the focus. The second
escape key press blurs the focus on the input box. The third escape key
closes the modal.

Though minor, this greatly improves the user experience of ordinary
keyboard navigation.