Improve tag list performance in large pages (#36)
* Improve tag list performance in large pages

This implements a virtualized viewport for the tag list and also makes
tag list calculation asynchronous.

Previously, the tag list used React deferred values to make rendering a
bit smoother in large pages, but one crucial issue with this is that the
value calculation can take a substantial amount of time in large pages
and this portion of work is synchronous. The new approach asynchronously
updates the tag list, with artificial yields inserted. Using something
like web workers would actually be ideal, but shipping the sometimes
huge metadata across the web workers boundary would probably be slow,
too. This approach works well enough.

In order to improve rendering with thousands of tags, the tag list
viewport is also virtualized. This alleviates the need to use deferred
rendering, since now the rendering is very quick.

The tag list also implements rendering a bit outside of the visible
viewport, using a value called "scroll slack". This is added to the page
view as well. This improves behavior when scrolling by relatively small
amounts.

* Virtual tag list improvements