Tweak ImageViewer image a bit
This alleviates some usability issues plus part of #9. There was intent
to make it so the transform transition gets reset when doing initial
positioning, but a reliable approach for this was never found. The
transition on transform didn't really seem to meaningfully work anyway,
so it may as well just not be there. User select is also disabled; you
could wind up selecting the image when dragging left/right with a mouse.
Finally, the image is made non-draggable.