Resolve hashes to file IDs for job view
I went through pains to allow pages that natively have hashes instead of
file IDs, but this is a pain in the ass all around. Most of the
functions that need to open pages with hashes are probably not going to
have over 10,000 files in them, which means it's really not going to
take an appreciable amount of time to just resolve the hashes to file
IDs eagerly, though it is definitely not ideal. On the bright side, this
will allow loading the *other* metadata using the cool new incremental
loading, so maybe the UX will be overall better. We really need file IDs
to render partial views, so this is probably the best we can do without
adding more caching.