Partial pages (#40)
* Implement partial page rendering

* Simplify page state slightly.

This fixes at least one bug where `loadedFiles` and `files` could fall
out of sync by simply getting rid of `files`. Now `fileIds` is used
anywhere where sparse files were sufficient.

This also will slightly improve performance of certain operations.

* Tweak ImageViewer image a bit

This alleviates some usability issues plus part of #9. There was intent
to make it so the transform transition gets reset when doing initial
positioning, but a reliable approach for this was never found. The
transition on transform didn't really seem to meaningfully work anyway,
so it may as well just not be there. User select is also disabled; you
could wind up selecting the image when dragging left/right with a mouse.
Finally, the image is made non-draggable.

* Prevent metadata updates from getting trampled during page load

* More fixes.

* Remove default API URL so Hydrus Companion will enter it.

* Handle trailing slash in API URL

* Allow the actual search part of a search to be cancelled

* Implement missing GetFileRelationships endpoint

* Resolve hashes to file IDs for job view

I went through pains to allow pages that natively have hashes instead of
file IDs, but this is a pain in the ass all around. Most of the
functions that need to open pages with hashes are probably not going to
have over 10,000 files in them, which means it's really not going to
take an appreciable amount of time to just resolve the hashes to file
IDs eagerly, though it is definitely not ideal. On the bright side, this
will allow loading the *other* metadata using the cool new incremental
loading, so maybe the UX will be overall better. We really need file IDs
to render partial views, so this is probably the best we can do without
adding more caching.

* Actually implement "best files"

This feature was in the UI but the implementation was unfinished! Oops.
It was one of the users of the hash virtual pages, so let's move it over
to file IDs.

* Remove virtual page support for hashes

* Fix bug preventing autosearch from working

* Fix incorrect state during search result load

* Get rid of old Tailwind cruft

* Make it possible to pause loading metadata

* Add metadata load state for FileViewerModal

* setIsLoadingPaused should perform wakeup

* Add partial state for TagList

* Update TagList.tsx

* Add eager loading preference