Detect Ladybird's stubs to prevent using OffscreenCanvas when it won't work
Ladybird probably stubs OffscreenCanvas so that apps that depend on it
continue to function even if it isn't critical for use. However, we are
detecting OffscreenCanvas support so we can fall back if it's not
supported, and Ladybird's stubs make the PSD viewer not work. Let's make
a quick patch over that.