FileViewer: Support multiple different viewers per file type (#43)
* Add /hydrui-server to .gitignore

* Do not use remote file for non-webworker PSD rendering either

Turns out the issue is with remote file, not with passing a blob to a
webworker.

* Detect Ladybird's stubs to prevent using OffscreenCanvas when it won't
work

Ladybird probably stubs OffscreenCanvas so that apps that depend on it
continue to function even if it isn't critical for use. However, we are
detecting OffscreenCanvas support so we can fall back if it's not
supported, and Ladybird's stubs make the PSD viewer not work. Let's make
a quick patch over that.

* Refactor FileViewer to allow multiple viewers per file

* A couple last-minute fixes.

* Fix blob URL leak in merged PSD renderer

* Uncommitted changes :')

* Copy tweaks