More UX fixes, some Hydrus features. (#57)
* TagLabel should ignore dash at beginning for namespace color

* Stop using HydrusApiClient interface

It is no longer needed.

* Add methods for (un)archiving, (un)deleting

* Disable "Remove files" in Hydrus tabs

It doesn't (and can't) work, so just disable it for now.

* Add archive, delete functionality to PageView

* Handle negated tags better

I've noticed Hydrus doesn't seem to support them for system tags, but
since it parses them it is still allowed.

* Slightly reduce padding in menus

This makes the touch target slightly smaller on touch devices, but the
menu is getting quite long, so it is probably overdue.

* Fix wildly incorrect durations

Hydrus uses milliseconds, not seconds. It's documented. This is probably
one of the longest-running Hydrui bugs.

* Add Deleted/Trashed/Inbox/Archived status to FilePreview

* Disable tab close button for Hydrus tabs

* Add demoserver support for archive/delete and tag negation

* Grammar fix

* Remove unused import